Abstract auth_service and extract implementation into a separate component
