+++ lib/mock-db/src/lib.rs
@@ -1,8 +1,10 @@
 use std::{collections::HashMap, sync::{Arc, Mutex, PoisonError}};
 
 use pheidippides_messenger::{Message, MessageId, User, UserId};
-use pheidippides_messenger::authorization::{AuthenticationInfo, AuthService, AuthStorage};
+use pheidippides_messenger::authorization::AuthService;
 use pheidippides_messenger::data_access::*;
+
+use pheidippides_auth::{AuthServiceImpl, AuthenticationInfo, AuthStorage};
 
 struct MessageRecord {
     id: MessageId,
@@ -114,7 +116,7 @@ impl Db {
         
         let res = Db { users, messages, auth };
 
-        let auth_service = AuthService::new(res.clone());
+        let auth_service = AuthServiceImpl::new(res.clone());
 
         let credentials = [
             (username_id_map["User1"], "User1"),
