+++ lib/pheidippides-auth/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "pheidippides-auth"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+pheidippides-messenger = {path="../pheidippides-messenger"}
+anyhow = "1.0.83"
+rand_core = "0.9.0-alpha.1"
+thiserror = "1.0.60"
+tokio = "1.37.0"
+password-hash = "0.5.0"
+argon2 = "0.5.3"
