+++ lib/pheidippides-messenger/Cargo.toml
@@ -7,9 +7,7 @@ edition = "2021"
 pheidippides-utils = {path= "../pheidippides-utils" }
 
 anyhow = "1.0.83"
-argon2 = "0.5.3"
 chrono = "0.4.38"
-password-hash = "0.5.0"
 rand_core = { version = "0.6.4", features = ["getrandom"] }
 thiserror = "1.0.60"
 tokio = { version = "1.37.0", features = ["time"] }
