+++ lib/pheidippides-web/src/request_handler.rs
@@ -1,6 +1,6 @@
 use pheidippides_messenger::data_access::DataAccess;
 use tokio::io::AsyncRead;
-use pheidippides_messenger::authorization::AuthStorage;
+use pheidippides_messenger::authorization::AuthService;
 use pheidippides_messenger::messenger::Messenger;
 use web_server::{Request, Response};
 use crate::routing;
@@ -35,7 +35,7 @@ impl std::fmt::Display for RequestHandlerError {
 
 impl std::error::Error for RequestHandlerError {}
 
-impl<D: DataAccess, A: AuthStorage, T: AsyncRead + Unpin + Sync + Send> web_server::RequestHandler<Request<T>> for RequestHandler<D, A> {
+impl<D: DataAccess, A: AuthService, T: AsyncRead + Unpin + Sync + Send> web_server::RequestHandler<Request<T>> for RequestHandler<D, A> {
     type Error = RequestHandlerError;
 
     fn handle(self, request: &mut Request<T>) -> impl std::future::Future<Output = anyhow::Result<Response, Self::Error>> + Send {
