+++ tests/Cargo.toml
@@ -7,6 +7,7 @@ publish = false
 [dev-dependencies]
 pheidippides-messenger = { path = "../lib/pheidippides-messenger" }
 pheidippides-web = { path = "../lib/pheidippides-web" }
+pheidippides-auth = { path = "../lib/pheidippides-auth" }
 web-server = { path = "../lib/web-server" }
 mock-db = { path = "../lib/mock-db" }
 postgres-db = { path = "../lib/postgres-db" }
